pyproject.toml
[project]
name = "chatyournotes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.16",
    "faiss-cpu>=1.11.0.post1",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.29",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.1.0",
    "tiktoken>=0.11.0",
]
